requirements.txt
# Core
torch>=2.3.0
torchvision>=0.18.0
diffusers>=0.32.0
transformers>=4.45.0
accelerate>=0.33.0
peft>=0.12.0

# Model loading & inference
safetensors>=0.4.3
huggingface_hub>=0.24.0
sentencepiece>=0.2.0

# Image processing
Pillow>=10.3.0
numpy>=1.26.0

# Captioning (WD14)
onnxruntime>=1.18.0

# Quantization
bitsandbytes>=0.43.0

# Training utilities
tensorboard>=2.17.0
tqdm>=4.66.0

# Dataset / captions
pandas>=2.2.0
